+++ packages/vue-oxlint-toolkit/js/index.ts
@@ -1,4 +1,4 @@
 export { parse } from './parse'
 export { transformJsx } from './transform-jsx'
-export type { ParseResult } from './parse-result'
-export type { Mapping, ToolkitTransformResult } from './transform-result'
+export type { ParseResult } from './parse'
+export type { Mapping, ToolkitTransformResult } from './transform-jsx'
