--- packages/vue-oxlint-toolkit/js/transform-result.ts
@@ -1,17 +1,0 @@
-import type { Comment, Diagnostic, Range } from '@oxlint/plugins'
-
-export interface Mapping {
-  virtualStart: number
-  virtualEnd: number
-  originalStart: number
-  originalEnd: number
-}
-
-export interface ToolkitTransformResult {
-  sourceText: string
-  scriptKind: 'jsx' | 'tsx'
-  comments: Comment[]
-  irregularWhitespaces: Range[]
-  errors: Diagnostic[]
-  mappings: Mapping[]
-}
